move common_elements util into unit
